Fix T22
This ended up being to relative paths verses absolute paths. I'm not
that happy with this; there's nothing I can put my finger on but I get
a feeling it's not the right way to it. Time will tell!

Because this is so confusing I've added some documentation to help the
next intrepid developer.